+++ src/plugin/PluginProcessor.h
@@ -20,6 +20,7 @@ public:
     static constexpr auto pickBiteParameterId = "pickBite";
     static constexpr auto pickStrokeParameterId = "pickStroke";
     static constexpr auto strumSpeedParameterId = "strumSpeed";
+    static constexpr auto strumBalanceParameterId = "strumBalance";
     static constexpr auto playerFeelParameterId = "playerFeel";
     static constexpr auto playerFeelRecoveryParameterId = "playerFeelRecovery";
     static constexpr auto palmMuteParameterId = "palmMute";
@@ -98,6 +99,7 @@ private:
     std::atomic<float>* pickBiteParameter = nullptr;
     std::atomic<float>* pickStrokeParameter = nullptr;
     std::atomic<float>* strumSpeedParameter = nullptr;
+    std::atomic<float>* strumBalanceParameter = nullptr;
     std::atomic<float>* playerFeelParameter = nullptr;
     std::atomic<float>* playerFeelRecoveryParameter = nullptr;
     std::atomic<float>* palmMuteParameter = nullptr;